feat: update telemetry transport to support Salute Eye API
